requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1725267563793/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560551097/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1726613473834/work
Flask==2.0.3
Flask-WTF==0.15.1
fonttools==4.56.0
frozenlist==1.4.1
fsspec==2024.9.0
gmpy2 @ file:///Users/runner/miniforge3/conda-bld/gmpy2_1725379799404/work
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1726459485162/work
itsdangerous==2.2.0
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1715127149914/work
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain-community==0.0.34
langchain-core==0.1.46
langchain-text-splitters==0.0.1
langsmith==0.1.129
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1724959500379/work
marshmallow==3.22.0
matplotlib==3.10.0
mlx==0.17.3
mlx-lm==0.18.2
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
multidict==6.1.0
mypy-extensions==1.0.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1712540363324/work
numpy==1.26.4
orjson==3.10.7
packaging==23.2
pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1726075093669/work
protobuf==5.28.2
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
pyparsing==3.2.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1725456256847/work
regex==2024.9.11
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
safetensors==0.4.5
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sympy==1.13.1
tenacity==8.5.0
tokenizers==0.19.1
torch==2.6.0.dev20240927
torchaudio==2.5.0.dev20240930
torchvision==0.20.0.dev20240930
tqdm==4.66.5
transformers==4.44.2
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1726496430923/work
Werkzeug==2.0.3
WTForms==3.1.2
yarl==1.13.1
zstandard==0.23.0
